refactor(strands): dedupe NAMS settings construction, promote AsyncBridge to integrations.base, close docs-import validation hole
